+++ libheif/context.cc
@@ -1434,7 +1434,7 @@ Result<std::shared_ptr<HeifPixelImage>> HeifContext::convert_to_output_colorspac
   bool different_chroma = (target_chroma != img->get_chroma_format());
   bool different_colorspace = (target_colorspace != img->get_colorspace());
 
-  uint8_t img_bpp = img->get_visual_image_bits_per_pixel();
+  uint16_t img_bpp = img->get_visual_image_bits_per_pixel();
   uint8_t converted_output_bpp = (options.convert_hdr_to_8bit && img_bpp > 8) ? 8 : 0 /* keep input depth */;
 
   nclx_profile img_nclx = img->get_color_profile_nclx_with_fallback();
